// src/app/shared/interfaces/additional.interface.ts
export interface IAdditional {
    'CREATE_DATE': string;
    'REQUEST_ID': string;
    'VEHICLE_MAKE': string;
    'MODELYEAR': string;
    'REG_MONTH': string;
    'REG_YEAR': string;
    'ODOMETER': string;
    'COLOUR': string;
    'PAINTCODE': string;
    'MODELSPEC': string;
    'CREATORSITEID': string;
    'AUDAUPDATEDBY': string;
    'AUDAVERSION': string;
    'SUBMITTEDBY': string;
    'DATEOFACCIDENT': string;
    'ASSESSOR_NAME': string;
    'ASSESSOR_ADDRESS': string;
}