// src/app/shared/interfaces/authorization_part.interface.ts
export interface IAuthorizationPart {
    APPROVED_DATE:  string;
    REQUEST_ID:  string;
    MATCHED_PART:  string;
    MATCHED_PART_ID:  string;
    AUDA_PART_ID:  string;
    BID_SUBMIT:  string;
    BID_SUBMIT_DATE:  string;
    WORKPROVIDER_APPROVED:  string;
    AUTHORISED_DATE:  string;
    WORKPROVIDER_AUTHORISED:  string;
    PART_DESCRIPTION: string;
    APPROVED_BY:  string;
    PART_PRICE: number;
    PART_ID: string;
    __v: number;
    _id: string;
}