// src/app/shared/interfaces/matched_part.interface.ts
export interface IMatchedPart {
    AUDA_PART_ID: string;
    CREATED_BY: string;
    CREATE_DATE: string;
    MATCHED_PART: string;
    MATCHED_PRICE: number;
    MATCH_DATE: string;
    PART_STANDARD: number;
    PART_STATUS_ID: number;
    QA_CONFIRMED: string;
    REQUEST_ID: string;
    PART_DESCRIPTION: string;
    STATUS: string;
    PART_ID: number;
    UPDATED_BY: string;
    UPDATED_DATE: string;
    __v: number;
    _id: string;
}