// src/app/shared/interfaces/user.interface.ts
export interface IUser {
        _id: string;
        createdBy: string;
        updatedBy: string;
        msisdn: string;
        firstName: string;
        lastName: string;
        updated: string;
        created: string;
        roles: any[];
        branch: string;
        insurer: string;
        language: string;
        STATUS: string;
        username: string;
  }